Example data of a weather station located in Caldern, Hessen, Germany.

A dataset containing different weather station measurements over the course of one day (2018-07-28). The weather station is located near Caldern, Hessen, Germany. Latitude: 50.840503 Longitude: 8.683300 Meters above sea level (elevation): 270 Heights of measurements: 2m (lower level) & 10m (upper level) Surface type: "Wiese" Ground texture: "clay"

weather_station_example_data

Format

A data frame with 288 rows and 20 variables:

datetime

Date and time of measurement in POSIXt-format.

t1

Temperature in °C at 2m height (lower level).

hum1

Humidity in % at 2m height (lower level).

t2

Temperature in °C at 10m height (upper level).

hum2

Humidity in % at 10m height (upper level).

rad_sw_in

Incoming shortwave radiation in W/m^2.

rad_sw_out

Outgoing shortwave radiation in W/m^2.

rad_lw_in

Incoming longwave radiation in W/m^2.

rad_lw_out

Outgoing longwave radiation in W/m^2.

rad_sw_bal

Shortwave radiation balance in W/m^2.

rad_lw_bal

Longwave radiation balance in W/m^2.

albedo

Albedo of ground.

rad_bal

Total radiation balance in W/m^2.

water_vol_soil

Moisture of ground in %.

t_surface

Surface temperature in °C.

ts1

Temperature Soil, 30 cm depth.

heatflux_soil

Soil heat flux in W/m^2.

v1

Wind velocity in m/s at 2m height (lower level).

v2

Wind velocity in m/s at 10m height (upper level).

p1

Air pressure in hPa (lower level).

p2

Air pressure in hPa (upper level).

Source

Provided by Prof. Dr. Jörg Bendix, Laboratory of Climatology and Remote Sensing, Philipps-University of Marburg.