fieldClim: Calculates Fluxes for Weather Stations

`fieldClim` is designed as a handy tool for calculations of various weather and micromicrometeorological parameters based on the measurements of a weather station. It includes functions adressing the estimation of radiation properties, the calculations of latent, sensible, and turbulent heat fluxes, the calculation of soil heat fluxes as well as the estimation of thermal and mechanical boundary layers.

Author

Maintainer: Fabian Mitze Mitze@students.uni-marburg.de

Authors:

Fabian Mitze Mitze@students.uni-marburg.de

David Langenohl langenoh@students.uni-marburg.de

Jonas Viehweger jonas.viehweger@rwth-aachen.de

Jörg Bendix bendix@geo.uni-marburg.de

Hanna Wenzel wenzel@students.uni-marburg.de

Kai Richter richter5@students.uni-marburg.de

You-De Chen you-de.chen@students.uni-marburg.de